pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FullIsingModel"
version = "1.0.0"
description = "Ising model components for quantum annealing workflows."
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["ising", "quantum", "annealing", "dwave", "pytorch"]

# Core runtime dependencies
dependencies = [
  "torch>=2.0",
  "numpy>=1.24",
  "dimod>=0.12.0",
  "dwave-neal>=0.6.0",
  "networkx>=3.0",
]

[project.optional-dependencies]
# Required only for the D-Wave QPU backend
qpu = ["dwave-system>=1.30.0"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["full_ising_model"]
